Remade config for Pydantic models
